requirements.txt
appdirs==1.4.4
lxml==6.0.1
packaging==25.0
pyshark==0.6
python-dotenv==1.1.1
termcolor==3.1.0
